finished assignment 2

+++ Bill3.java
@@ -5,7 +5,9 @@ public class Bill3 {
 	    String name1 = args[0];
 	    String name2 = args[1];
 		String name3 = args[2];
-		double price = Math.ceil((double)Integer.parseInt(args[3])/3); 
+		double price = Math.ceil(Double.parseDouble(args[3])/3); 
 		System.out.println("Dear " + name3 + ", " + name2 + ", and " + name1 + ": pay " + price + " Shekels each");
+		//double price = Math.ceil((double)Integer.parseInt(args[3])/3); - incase the price line doesnt work after the change
 	}
 }
+
